Added jax gaussian aug. and restructured package.
Jax based augmented is added, but in order to avoid circular
dependencies.

Trajectories are not yet inculpated into map generation.
