+++ src/aggforce/__init__.py
@@ -5,7 +5,17 @@ resolution simulation. This process requires the higher resolution forces be map
 This module provides routines for performing this mapping in different ways, but does
 not itself parameterize any force-fields.
 
-The primary entry point is agg.project_forces.
+The primary entry point is project_forces. Only basic routines for simple linear map 
+optimization are visible in the module namespace; more advanced features require 
+explicit imports from submodules.
 
 See agg.py and the README for more information. Tests and examples are also available.
 """
+
+# __init__ doesn't use the imported objects
+# ruff: noqa: F401
+from .trajectory import Trajectory
+from .agg import project_forces
+from .constraints import guess_pairwise_constraints
+from .qp import qp_linear_map, constraint_aware_uni_map
+from .map import LinearMap
