+++ tests/test_forces.py
@@ -42,9 +42,12 @@ import numpy as np
 import numpy.random as r
 import mdtraj as md  # type: ignore [import-untyped]
 import pytest
-from aggforce import agg as ag
-from aggforce import linearmap as lm
-from aggforce import constfinder as cf
+from aggforce import (
+    guess_pairwise_constraints,
+    LinearMap,
+    project_forces,
+    constraint_aware_uni_map,
+)
 
 
 # this seeds some portions of the randomness of these tests, but not be
@@ -93,7 +96,7 @@ def get_data() -> Tuple[np.ndarray, np.ndarray, md.Trajectory, float]:
     return (coords, forces, pdb, kbt)
 
 
-def gen_config_map(pdb: md.Trajectory, string: str = "CA$") -> lm.LinearMap:
+def gen_config_map(pdb: md.Trajectory, string: str = "CA$") -> LinearMap:
     r"""Create the configurational map.
 
     This is needed as it defines constraints which dictate which force maps are
@@ -122,7 +125,7 @@ def gen_config_map(pdb: md.Trajectory, string: str = "CA$") -> lm.LinearMap:
     for ind, a in enumerate(atomlist):
         if re.search(string, str(a)):
             inds.append([ind])
-    return lm.LinearMap(inds, n_fg_sites=pdb.xyz.shape[1])
+    return LinearMap(inds, n_fg_sites=pdb.xyz.shape[1])
 
 
 def test_cln025_basic_agg_forces_against_ref() -> None:
@@ -135,15 +138,15 @@ def test_cln025_basic_agg_forces_against_ref() -> None:
     # cmap is the configurational coarse-grained map
     cmap = gen_config_map(pdb, "CA$")
     # guess molecular constraints
-    constraints = cf.guess_pairwise_constraints(coords[0:10], threshold=1e-3)
+    constraints = guess_pairwise_constraints(coords[0:10], threshold=1e-3)
 
     # make force map
-    basic_results = ag.project_forces(
+    basic_results = project_forces(
         xyz=coords,
         forces=forces,
         config_mapping=cmap,
         constrained_inds=constraints,
-        method=lm.constraint_aware_uni_map,
+        method=constraint_aware_uni_map,
     )
 
     location = Path(__file__).parent
@@ -162,10 +165,10 @@ def test_cln025_opt_agg_forces_against_ref() -> None:
     # cmap is the configurational coarse-grained map
     cmap = gen_config_map(pdb, "CA$")
     # guess molecular constraints
-    constraints = cf.guess_pairwise_constraints(coords[0:10], threshold=1e-3)
+    constraints = guess_pairwise_constraints(coords[0:10], threshold=1e-3)
 
     # make force map
-    optim_results = ag.project_forces(
+    optim_results = project_forces(
         xyz=coords,
         forces=forces,
         config_mapping=cmap,
@@ -222,7 +225,7 @@ def test_cln025_opt_basic_rsqpg_mscg_ip(seed: int = rseed) -> None:
     # cmap is the configurational coarse-grained map
     cmap = gen_config_map(pdb, "CA$")
     # guess molecular constraints
-    constraints = cf.guess_pairwise_constraints(coords[0:10], threshold=1e-3)
+    constraints = guess_pairwise_constraints(coords[0:10], threshold=1e-3)
 
     train_coords = coords[:500]
     test_coords = coords[500:]
@@ -230,15 +233,15 @@ def test_cln025_opt_basic_rsqpg_mscg_ip(seed: int = rseed) -> None:
     train_forces = forces[:500]
     test_forces = forces[500:]
 
-    basic_results = ag.project_forces(
+    basic_results = project_forces(
         xyz=train_coords,
         forces=train_forces,
         config_mapping=cmap,
         constrained_inds=constraints,
-        method=lm.constraint_aware_uni_map,
+        method=constraint_aware_uni_map,
     )
 
-    optim_results = ag.project_forces(
+    optim_results = project_forces(
         xyz=train_coords,
         forces=train_forces,
         config_mapping=cmap,
@@ -327,7 +330,7 @@ def test_cln025_opt_basic_rsqpg_offset(seed: int = rseed) -> None:
     # cmap is the configurational coarse-grained map
     cmap = gen_config_map(pdb, "CA$")
     # guess molecular constraints
-    constraints = cf.guess_pairwise_constraints(coords[0:10], threshold=1e-3)
+    constraints = guess_pairwise_constraints(coords[0:10], threshold=1e-3)
 
     train_coords = coords[:500]
     test_coords = coords[500:]
@@ -335,15 +338,15 @@ def test_cln025_opt_basic_rsqpg_offset(seed: int = rseed) -> None:
     train_forces = forces[:500]
     test_forces = forces[500:]
 
-    basic_results = ag.project_forces(
+    basic_results = project_forces(
         xyz=train_coords,
         forces=train_forces,
         config_mapping=cmap,
         constrained_inds=constraints,
-        method=lm.constraint_aware_uni_map,
+        method=constraint_aware_uni_map,
     )
 
-    optim_results = ag.project_forces(
+    optim_results = project_forces(
         xyz=train_coords,
         forces=train_forces,
         config_mapping=cmap,
@@ -403,14 +406,20 @@ def test_cln025_featopt_opt_rsqpg_mscg_ip(seed: int = rseed) -> None:
     marginal gain in the force score.
     """
     from aggforce import jaxmapval as mv
-    from aggforce import featlinearmap as p
-    from aggforce import jaxfeat as jf
+    from aggforce.qp import (
+        Multifeaturize,
+        gb_feat,
+        GeneralizedFeatures,
+        id_feat,
+        qp_feat_linear_map,
+    )
+    from aggforce.util import Curry
 
     coords, forces, pdb, kbt = get_data()
     # cmap is the configurational coarse-grained map
     cmap = gen_config_map(pdb, "CA$")
     # guess molecular constraints
-    constraints = cf.guess_pairwise_constraints(coords[0:10], threshold=1e-3)
+    constraints = guess_pairwise_constraints(coords[0:10], threshold=1e-3)
 
     N_FRAMES: Final = 500
 
@@ -420,7 +429,7 @@ def test_cln025_featopt_opt_rsqpg_mscg_ip(seed: int = rseed) -> None:
     train_forces = forces[:N_FRAMES]
     test_forces = forces[N_FRAMES:]
 
-    basic_optim_results = ag.project_forces(
+    basic_optim_results = project_forces(
         xyz=train_coords,
         forces=train_forces,
         config_mapping=cmap,
@@ -428,22 +437,22 @@ def test_cln025_featopt_opt_rsqpg_mscg_ip(seed: int = rseed) -> None:
         l2_regularization=1e3,
     )
 
-    f0: p.Curry[p.GeneralizedFeatures] = p.Curry(
-        jf.gb_feat,
+    f0: Curry[GeneralizedFeatures] = Curry(
+        gb_feat,
         inner=0.0,
         outer=8.0,
         width=1.0,
         n_basis=4,
         lazy=True,
     )
-    comb_f = p.Multifeaturize([p.id_feat, f0])  # type: ignore [list-item]
-    optim_results = ag.project_forces(
+    comb_f = Multifeaturize([id_feat, f0])  # type: ignore [list-item]
+    optim_results = project_forces(
         xyz=coords,
         forces=forces,
         config_mapping=cmap,
         constrained_inds=constraints,
         featurizer=comb_f,
-        method=p.qp_feat_linear_map,
+        method=qp_feat_linear_map,
         kbt=kbt,
         l2_regularization=1e4,
     )
@@ -501,14 +510,19 @@ def test_cln025_featopt_opt_rsqpg_offset(seed: int = rseed) -> None:
     marginal gain in the force score.
     """
     from aggforce import jaxmapval as mv
-    from aggforce import featlinearmap as p
-    from aggforce import jaxfeat as jf
+    from aggforce.qp import (
+        Multifeaturize,
+        gb_feat,
+        id_feat,
+        qp_feat_linear_map,
+    )
+    from aggforce.util import Curry
 
     coords, forces, pdb, kbt = get_data()
     # cmap is the configurational coarse-grained map
     cmap = gen_config_map(pdb, "CA$")
     # guess molecular constraints
-    constraints = cf.guess_pairwise_constraints(coords[0:10], threshold=1e-3)
+    constraints = guess_pairwise_constraints(coords[0:10], threshold=1e-3)
 
     train_coords = coords[:500]
     test_coords = coords[500:]
@@ -516,29 +530,29 @@ def test_cln025_featopt_opt_rsqpg_offset(seed: int = rseed) -> None:
     train_forces = forces[:500]
     test_forces = forces[500:]
 
-    basic_opt_results = ag.project_forces(
+    basic_opt_results = project_forces(
         xyz=train_coords,
         forces=train_forces,
         config_mapping=cmap,
         constrained_inds=constraints,
-        method=lm.constraint_aware_uni_map,
+        method=constraint_aware_uni_map,
     )
-    f0 = p.Curry(
-        jf.gb_feat,
+    f0 = Curry(
+        gb_feat,
         inner=0.0,
         outer=8.0,
         width=1.0,
         n_basis=4,
         lazy=True,
     )
-    comb_f = p.Multifeaturize([p.id_feat, f0])  # type: ignore [list-item]
-    optim_results = ag.project_forces(
+    comb_f = Multifeaturize([id_feat, f0])  # type: ignore [list-item]
+    optim_results = project_forces(
         xyz=coords,
         forces=forces,
         config_mapping=cmap,
         constrained_inds=constraints,
         featurizer=comb_f,
-        method=p.qp_feat_linear_map,
+        method=qp_feat_linear_map,
         kbt=kbt,
         l2_regularization=1e4,
     )
